requirements.txt
click==8.0.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
PyMySQL==0.10.0
Werkzeug==2.0.2
inquirer~=2.8.0
pandas~=1.3.4
numpy~=1.21.4